+++ bundles/org.eclipse.swt/Eclipse SWT/gtk/org/eclipse/swt/widgets/DirectoryDialog.java
@@ -154,9 +154,9 @@ public String open () {
 	if (response == OS.GTK_RESPONSE_OK) {
 		int /*long*/ fileNamePtr = OS.gtk_file_selection_get_filename (handle);
 		int /*long*/ utf8Ptr = OS.g_filename_to_utf8 (fileNamePtr, -1, null, null, null);
-		int [] items_written = new int [1];
+		int /*long*/ [] items_written = new int /*long*/ [1];
 		int /*long*/ utf16Ptr = OS.g_utf8_to_utf16 (utf8Ptr, -1, null, items_written, null);
-		int length = items_written [0];
+		int length = (int)/*64*/items_written [0];
 		char [] buffer = new char [length];
 		OS.memmove (buffer, utf16Ptr, length * 2);
 		String osAnswer = new String (buffer);
