+++ bundles/org.eclipse.swt/Eclipse SWT/gtk/org/eclipse/swt/widgets/FileDialog.java
@@ -170,9 +170,9 @@ public String open () {
 	if (response == OS.GTK_RESPONSE_OK) {
 		int /*long*/ fileNamePtr = OS.gtk_file_selection_get_filename (handle);
 		int /*long*/ utf8Ptr = OS.g_filename_to_utf8 (fileNamePtr, -1, null, null, null);
-		int [] items_written = new int [1];
+		int /*long*/ [] items_written = new int /*long*/ [1];
 		int /*long*/ utf16Ptr = OS.g_utf8_to_utf16 (utf8Ptr, -1, null, items_written, null);
-		int length = items_written [0];
+		int length = (int)/*64*/items_written [0];
 		char [] buffer = new char [length];
 		OS.memmove (buffer, utf16Ptr, length * 2);
 		String osAnswer = new String (buffer);
@@ -294,9 +294,9 @@ String interpretOsAnswer(String osAnswer) {
 		OS.memmove (namePtr, namesPtr, length * OS.PTR_SIZEOF);
 		for (int i = 0; i < length; i++) {			
 			int /*long*/ utf8Ptr = OS.g_filename_to_utf8 (namePtr [i], -1, null, null, null);
-			int [] items_written = new int [1];
+			int /*long*/ [] items_written = new int /*long*/ [1];
 			int /*long*/ utf16Ptr = OS.g_utf8_to_utf16 (utf8Ptr, -1, null, items_written, null);
-			char[] buffer = new char [items_written [0]];
+			char[] buffer = new char [(int)/*64*/items_written [0]];
 			OS.memmove (buffer, utf16Ptr, items_written [0] * 2);
 			String name = new String (buffer);
 			fileNames [i] = name.substring (name.lastIndexOf (SEPARATOR) + 1);
